Extract Hash.h, Hash.cpp header and implementation in audio/
